# backend-laravel/storage/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "D\u00e9panne-Moi",
        "version": "1.0.0",
        "description": "API de l'application mobile D\u00e9panne-Moi \u2014 mise en relation entre clients\net techniciens v\u00e9rifi\u00e9s \u00e0 Conakry.\n\n## Authentification\n\nToutes les routes prot\u00e9g\u00e9es attendent un en-t\u00eate\n`Authorization: Bearer <access_token>`. L'access token expire en\n**15 minutes** ; le jeton de rafra\u00eechissement, valable 30 jours, permet\nd'en obtenir un nouveau via `POST /api/v1/auth/refresh`.\n\nLe jeton de rafra\u00eechissement **tourne \u00e0 chaque usage** : la r\u00e9ponse en\nrenvoie toujours un nouveau, et l'ancien cesse imm\u00e9diatement d'\u00eatre\nvalable. Pr\u00e9senter deux fois le m\u00eame r\u00e9voque toutes les sessions du\ncompte \u2014 c'est le signe qu'une copie circule.\n\n## Conventions\n\n- Les num\u00e9ros de t\u00e9l\u00e9phone sont en E.164 (`+224XXXXXXXXX`) ; l'API\n  accepte les formes locales et les normalise.\n- Les montants sont des **entiers de francs guin\u00e9ens**, sans d\u00e9cimale.\n- Les horodatages sont en ISO 8601, fuseau `Africa/Conakry` (UTC+0).\n- Les messages d'erreur sont en fran\u00e7ais et destin\u00e9s \u00e0 \u00eatre affich\u00e9s\n  tels quels \u00e0 l'utilisateur."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "paths": {
...
        "/moi": {
            "get": {
                "operationId": "moi",
                "summary": "Mon compte, avec les profils correspondant \u00e0 mes casquettes",
                "tags": [
                    "Profil"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "utilisateur": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/UtilisateurResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "profil_client",
                                                        "profil_technicien"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "utilisateur"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "moi.modifier",
                "summary": "Modifier mes informations. Le t\u00e9l\u00e9phone n'est pas modifiable ici : c'est l'identifiant de connexion",
                "tags": [
                    "Profil"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "minLength": 3,
                                        "maxLength": 120
                                    },
                                    "email": {
                                        "type": [
                                            "string",
                                            "null"
                                        ],
                                        "format": "email",
                                        "maxLength": 190
                                    },
                                    "avatar_url": {
                                        "type": [
                                            "string",
                                            "null"
                                        ],
                                        "maxLength": 500
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Informations mises \u00e0 jour."
                                        },
                                        "utilisateur": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/UtilisateurResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "profil_client",
                                                        "profil_technicien"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "utilisateur"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/auth/connexion": {
            "post": {
                "operationId": "connexion",
                "description": "Num\u00e9ro de t\u00e9l\u00e9phone et mot de passe \u2014 pas d'OTP. Cinq tentatives par\nfen\u00eatre de quinze minutes et par couple num\u00e9ro + IP.",
                "summary": "Se connecter",
                "tags": [
                    "Session"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConnexionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "utilisateur": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/UtilisateurResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "profil_client",
                                                        "profil_technicien"
                                                    ]
                                                }
                                            ]
                                        },
                                        "jetons": {
                                            "type": "object",
                                            "properties": {
                                                "access_token": {
                                                    "type": "string"
                                                },
                                                "refresh_token": {
                                                    "type": "string"
                                                },
                                                "expire_dans": {
                                                    "type": "string"
                                                },
                                                "expire_le": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "access_token",
                                                "refresh_token",
                                                "expire_dans",
                                                "expire_le"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "utilisateur",
                                        "jetons"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "operationId": "refresh",
                "description": "\u00c9change un jeton de rafra\u00eechissement contre un couple neuf. Le jeton\npr\u00e9sent\u00e9 est imm\u00e9diatement r\u00e9voqu\u00e9 : il ne sert qu'une fois.",
                "summary": "Rafra\u00eechir la session",
                "tags": [
                    "Session"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "refresh_token": {
                                        "type": "string"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "maxLength": 120
                                    }
                                },
                                "required": [
                                    "refresh_token"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "jetons": {
                                            "type": "object",
                                            "properties": {
                                                "access_token": {
                                                    "type": "string"
                                                },
                                                "refresh_token": {
                                                    "type": "string"
                                                },
                                                "expire_dans": {
                                                    "type": "string"
                                                },
                                                "expire_le": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "access_token",
                                                "refresh_token",
                                                "expire_dans",
                                                "expire_le"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "jetons"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ConnexionRequest": {
                "type": "object",
                "properties": {
                    "phone": {
                        "type": "string",
                        "maxLength": 20
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 72
                    },
                    "device_name": {
                        "type": "string",
                        "description": "Nom de l'appareil : il appara\u00eet dans la liste des sessions et\npermet au porteur de reconna\u00eetre laquelle r\u00e9voquer.",
                        "maxLength": 120
                    }
                },
                "required": [
                    "phone",
                    "password"
                ],
                "title": "ConnexionRequest"
            },
...
            "ProfilClientResource": {
                "type": "object",
                "properties": {
                    "points_fidelite": {
                        "type": "integer"
                    },
                    "interventions": {
                        "type": "integer"
                    },
                    "adresse_par_defaut_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    }
                },
                "required": [
                    "points_fidelite",
                    "interventions",
                    "adresse_par_defaut_id"
                ],
                "title": "ProfilClientResource"
            },
            "ProfilTechnicienResource": {
                "type": "object",
                "properties": {
                    "specialites": {
                        "type": "array",
                        "items": {}
                    },
                    "verification": {
                        "type": "object",
                        "properties": {
                            "statut": {
                                "type": "string"
                            },
                            "libelle": {
                                "type": "string",
                                "enum": [
                                    "En attente de validation",
                                    "Valid\u00e9",
                                    "Rejet\u00e9",
                                    "Suspendu"
                                ]
                            },
                            "peut_travailler": {
                                "type": "boolean"
                            },
                            "motif_rejet": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "validee_le": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        },
                        "required": [
                            "statut",
                            "libelle",
                            "peut_travailler",
                            "motif_rejet",
                            "validee_le"
                        ]
                    },
                    "zone": {
                        "type": "object",
                        "properties": {
                            "latitude": {
                                "type": "null"
                            },
                            "longitude": {
                                "type": "null"
                            },
                            "rayon_km": {
                                "type": "integer"
                            }
                        },
                        "required": [
                            "latitude",
                            "longitude",
                            "rayon_km"
                        ]
                    },
                    "en_ligne": {
                        "type": "boolean"
                    },
                    "statistiques": {
                        "type": "object",
                        "properties": {
                            "note_moyenne": {
                                "type": "number"
                            },
                            "note_appliquee": {
                                "type": "number"
                            },
                            "avis": {
                                "type": "integer"
                            },
                            "interventions": {
                                "type": "integer"
                            },
                            "taux_acceptation": {
                                "type": "number"
                            },
                            "taux_annulation": {
                                "type": "number"
                            },
                            "debutant": {
                                "type": "boolean"
                            }
                        },
                        "required": [
                            "note_moyenne",
                            "note_appliquee",
                            "avis",
                            "interventions",
                            "taux_acceptation",
                            "taux_annulation",
                            "debutant"
                        ]
                    }
                },
                "required": [
                    "specialites",
                    "verification",
                    "zone",
                    "en_ligne",
                    "statistiques"
                ],
                "title": "ProfilTechnicienResource"
            },
...
            "UtilisateurResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nom_complet": {
                        "type": "string"
                    },
                    "prenom": {
                        "type": "string"
                    },
                    "telephone": {
                        "type": "string"
                    },
                    "telephone_affichage": {
                        "type": "string"
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "avatar_url": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "statut": {
                        "type": "string"
                    },
                    "casquettes": {
                        "type": "object",
                        "properties": {
                            "client": {
                                "type": "boolean"
                            },
                            "technicien": {
                                "type": "boolean"
                            }
                        },
                        "required": [
                            "client",
                            "technicien"
                        ]
                    },
                    "profil_client": {
                        "$ref": "#/components/schemas/ProfilClientResource"
                    },
                    "profil_technicien": {
                        "$ref": "#/components/schemas/ProfilTechnicienResource"
                    },
                    "inscrit_le": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "derniere_connexion": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "nom_complet",
                    "prenom",
                    "telephone",
                    "telephone_affichage",
                    "email",
                    "avatar_url",
                    "statut",
                    "casquettes",
                    "inscrit_le",
                    "derniere_connexion"
                ],
                "title": "UtilisateurResource"
            }
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}